Fix CI: portable errno, Linux errno aliasing, Windows build and tests
- lock.rs: `libc::__error` exists only on macOS/BSD; read errno via
  `io::Error::last_os_error()` so the Linux build compiles.
- contain/unix.rs: ENOTSUP and EOPNOTSUPP are the same value on Linux,
  so the or-pattern was flagged unreachable under `-D warnings`. Use a
  match guard instead.
- contain/windows.rs: winapi-util 0.1.11 returns file attributes as u64;
  narrow to the u32 the FILE_ATTRIBUTE_* masks expect.
- cli/maintenance.rs: `problems` is only mutated on unix; allow the
  unused `mut` on other targets.
- restore/place.rs: the placement tests use unix modes and symlinks;
  gate the module on unix.

+++ src/restore/contain/windows.rs
@@ -47,7 +47,7 @@ fn open_reparse_aware(path: &Path, write: bool) -> io::Result<File> {
 fn attributes(file: &File) -> io::Result<(u32, u64, u64)> {
     let info = winapi_util::file::information(file)?;
     Ok((
-        info.file_attributes(),
+        info.file_attributes() as u32,
         info.volume_serial_number(),
         info.file_index(),
     ))
